requirements.txt
langgraph
fastapi
nova-act
langchain-aws
langgraph
langchain_community
pyjwt>=2.10.1
opentelemetry-instrumentation-langchain
bedrock-agentcore
bedrock-agentcore-starter-toolkit
duckduckgo-search